reformatted education section on intake form
