+++ src/styles/css/styles.css
@@ -935,4 +935,12 @@ td.ant-table-column-sort {
   color: #d2d2d7;
 }
 
+.education {
+  padding-bottom: 10rem;
+}
+
+.radioButtons {
+  padding-left: 1rem;
+}
+
